style: 상태 관련 클래스 패키지 변경
