+++ src/main/java/blackjack/domain/state/finished/Blackjack.java
@@ -1,16 +1,16 @@
-package blackjack.domain.state;
+package blackjack.domain.state.finished;
 
 import blackjack.domain.GivenCards;
 import blackjack.domain.enums.Result;
 
-public class Burst extends Finished {
+public class Blackjack extends Finished {
 
-    public Burst(final GivenCards givenCards) {
+    public Blackjack(final GivenCards givenCards) {
         super(givenCards);
     }
 
     @Override
     public int result(final GivenCards competitorCards) {
-        return Result.LOSE.getPoint();
+        return Result.WIN.getPoint();
     }
 }
