+++ src/main/java/blackjack/domain/state/finished/Burst.java
@@ -1,16 +1,16 @@
-package blackjack.domain.state;
+package blackjack.domain.state.finished;
 
 import blackjack.domain.GivenCards;
 import blackjack.domain.enums.Result;
 
-public class Blackjack extends Finished {
+public class Burst extends Finished {
 
-    public Blackjack(final GivenCards givenCards) {
+    public Burst(final GivenCards givenCards) {
         super(givenCards);
     }
 
     @Override
     public int result(final GivenCards competitorCards) {
-        return Result.WIN.getPoint();
+        return Result.LOSE.getPoint();
     }
 }
