+++ src/main/java/blackjack/domain/state/finished/Stay.java
@@ -1,4 +1,4 @@
-package blackjack.domain.state;
+package blackjack.domain.state.finished;
 
 import blackjack.domain.GivenCards;
 import blackjack.domain.enums.Result;
