+++ src/main/java/blackjack/domain/state/running/Running.java
@@ -1,6 +1,7 @@
-package blackjack.domain.state;
+package blackjack.domain.state.running;
 
 import blackjack.domain.GivenCards;
+import blackjack.domain.state.Init;
 
 public abstract class Running extends Init {
